+++ src/Pim/Bundle/CatalogBundle/AttributeType/NumberType.php
@@ -57,7 +57,7 @@ class NumberType extends FlexNumberType
                 'fieldType' => 'switch'
             ),
             array(
-                'name'      => 'translatable',
+                'name'      => 'localizable',
                 'fieldType' => 'switch',
                 'options'   => array(
                     'disabled'  => (bool) $attribute->getId(),