+++ src/Pim/Bundle/CatalogBundle/AttributeType/PriceCollectionType.php
@@ -81,7 +81,7 @@ class PriceCollectionType extends AbstractAttributeType
                 'fieldType' => 'switch'
             ),
             array(
-                'name'      => 'translatable',
+                'name'      => 'localizable',
                 'fieldType' => 'switch',
                 'options'   => array(
                     'disabled'  => (bool) $attribute->getId(),